+++ app/Http/Controllers/RoleController.php
@@ -7,18 +7,23 @@ use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
 use App\Models\Permission;
 use App\Models\PermissionRoleModel;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class RoleController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index()
     {
+        $this->authorize('view_roles');
         $roles = Role::latest()->get();
         return view('dashboard.admin.roles.index', compact('roles'));
     }
 
     public function create()
     {
+        $this->authorize('create_roles');
         $getPermission = Permission::getRecord();
         $data['getPermission'] = $getPermission;
         return view('dashboard.admin.roles.create', $data);
@@ -37,6 +42,7 @@ class RoleController extends Controller
 
     public function edit(string $id)
     {
+        $this->authorize('edit_roles');
         $data['getRecord'] = Role::getSingle($id);
         $data['getPermission'] = Permission::getRecord();
         $data['getRolePermission'] = PermissionRoleModel::getRolePermission($id);
@@ -55,6 +61,7 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        $this->authorize('delete_roles');
         // Hapus data permission_role yang terkait
         \App\Models\PermissionRoleModel::where('role_id', $role->id)->delete();
